Move shared logic for github handlers into base class
