+++ lib/txgh/handlers/github.rb
@@ -2,6 +2,7 @@ module Txgh
   module Handlers
     module Github
       autoload :DeleteHandler,  'txgh/handlers/github/delete_handler'
+      autoload :Handler,        'txgh/handlers/github/handler'
       autoload :PushHandler,    'txgh/handlers/github/push_handler'
       autoload :RequestHandler, 'txgh/handlers/github/request_handler'
     end
