+++ lib/txgh/handlers/github/delete_handler.rb
@@ -1,19 +1,7 @@
-require 'logger'
-
 module Txgh
   module Handlers
     module Github
-      class DeleteHandler
-        include ResponseHelpers
-
-        attr_reader :project, :repo, :payload, :logger
-
-        def initialize(options = {})
-          @project = options.fetch(:project)
-          @repo = options.fetch(:repo)
-          @payload = options.fetch(:payload)
-          @logger = options.fetch(:logger) { Logger.new(STDOUT) }
-        end
+      class DeleteHandler < Handler
 
         def execute
           if should_handle_request?
@@ -65,6 +53,7 @@ module Txgh
         def branch
           Utils.absolute_branch(payload['ref'].sub(/^refs\//, ''))
         end
+
       end
     end
   end
