+++ lib/txgh/handlers/github/handler.rb
@@ -1,0 +1,20 @@
+require 'logger'
+
+module Txgh
+  module Handlers
+    module Github
+      class Handler
+        include ResponseHelpers
+
+        attr_reader :project, :repo, :payload, :logger
+
+        def initialize(options = {})
+          @project = options.fetch(:project)
+          @repo = options.fetch(:repo)
+          @payload = options.fetch(:payload)
+          @logger = options.fetch(:logger) { Logger.new(STDOUT) }
+        end
+      end
+    end
+  end
+end
